增加收藏功能后

+++ src/com/sjzg/answer/AnswerGetByID.java
@@ -161,7 +161,7 @@ public class AnswerGetByID extends HttpServlet {
 
 
 
-				 answerQuestionModel_temp.setTimeUsed(rs.getInt("TimeUsed"));
+				 answerQuestionModel_temp.setTimeUsed(rs.getFloat("TimeUsed"));
 				 answerQuestionModel_temp.setTrack(rs.getString("Track"));
 				 answerQuestionModel_temp.setGrade(rs.getFloat("Grade"));
 				answerQuestionModel_temp.setUnfocusDuration(rs.getInt("UnfocusDuration"));
