+++ src/com/sjzg/rank/RankByMistake.java
@@ -265,7 +265,7 @@ public class RankByMistake extends HttpServlet {
 				
 				AnswerQuestionModel answerQuestionModel_temp = new AnswerQuestionModel();
 				 answerQuestionModel_temp.setQuestionID(rs.getInt("QuestionID"));
-				 answerQuestionModel_temp.setTimeUsed(rs.getInt("TimeUsed"));
+				 answerQuestionModel_temp.setTimeUsed(rs.getFloat("TimeUsed"));
 				 answerQuestionModel_temp.setTrack(rs.getString("Track"));
 				 answerQuestionModel_temp.setGrade(rs.getFloat("Grade"));
 				answerQuestionModel_temp.setUnfocusDuration(rs.getInt("UnfocusDuration"));
